REFS #5 by Berna-L: Created new Builder-based API with cascading Lambdas.
Based on the comments on https://javatechnicalwealth.com/blog/2018/04/16/cascading-lambdas-and-builder-pattern-in-java-when-1-1-3-but-not-4/
The only issue is that in-IDE Javadoc doesn't seem to work (probably because we are passing method references, not the methods themselves), but a visit to the HTML Javadoc should work. I'll keep this in the backburner for now.